feat(security): fix argument injection and lint issues
- Add `is_safe_url` helper in `slatron-server/src/rhai_engine/mod.rs` to validate URLs.
- Reject URLs starting with `-` to prevent `curl` argument injection.
- Enforce `http://` or `https://` schemes to prevent SSRF/LFI via `file://`.
- Add comprehensive tests in `slatron-server/src/rhai_engine/path_test.rs`.
- Fix existing clippy warnings and formatting issues in `slatron-server` to pass CI.

Sentinel Priority: CRITICAL (Argument Injection)

+++ slatron-server/src/rhai_engine/mod.rs
@@ -198,7 +198,7 @@ fn run_shell_execute(cmd: String, args: Vec<rhai::Dynamic>) -> rhai::Map {
             let err_msg = format!("Execution failed: {}", e);
             tracing::error!("Shell Spawn Failed: {}", err_msg);
 
-            map.insert("code".into(), (-1 as i64).into());
+            map.insert("code".into(), (-1_i64).into());
             map.insert("stdout".into(), "".into());
             map.insert("stderr".into(), err_msg.into());
         }
